Refactor utility functions: group into a single functions module
