+++ operators/gap_remove.py
@@ -1,7 +1,7 @@
 import bpy
 from operator import attrgetter
 
-from .utils.slice_contiguous_sequence_list import slice_selection
+from .utils.functions import slice_selection
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
