+++ operators/jump_time_offset.py
@@ -1,6 +1,6 @@
 import bpy
 
-from .utils.convert_duration_to_frames import convert_duration_to_frames
+from .utils.functions import convert_duration_to_frames
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
