+++ operators/make_still_image.py
@@ -2,7 +2,7 @@ import bpy
 import operator
 
 from .utils.global_settings import SequenceTypes
-from .utils.convert_duration_to_frames import convert_duration_to_frames
+from .utils.functions import convert_duration_to_frames
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
