+++ operators/preview_closest_cut.py
@@ -1,8 +1,8 @@
 import bpy
 
-from .utils.get_frame_range import get_frame_range
-from .utils.set_preview_range import set_preview_range
-from .utils.convert_duration_to_frames import convert_duration_to_frames
+from .utils.functions import get_frame_range
+from .utils.functions import set_preview_range
+from .utils.functions import convert_duration_to_frames
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
