+++ operators/preview_to_selection.py
@@ -1,7 +1,7 @@
 import bpy
 
-from .utils.get_frame_range import get_frame_range
-from .utils.set_preview_range import set_preview_range
+from .utils.functions import get_frame_range
+from .utils.functions import set_preview_range
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
