+++ operators/select_closest_to_mouse.py
@@ -1,7 +1,7 @@
 import bpy
 
-from .utils.find_strips_mouse import find_strips_mouse
-from .utils.get_mouse_view_coords import get_mouse_frame_and_channel
+from .utils.functions import find_strips_mouse
+from .utils.functions import get_mouse_frame_and_channel
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
