+++ operators/trim_three_point_edit.py
@@ -1,6 +1,6 @@
 import bpy
 
-from .utils.get_mouse_view_coords import get_mouse_frame_and_channel
+from .utils.functions import get_mouse_frame_and_channel
 from .utils.doc import doc_name, doc_idname, doc_brief, doc_description
