--- operators/utils/calculate_distance.py
@@ -1,5 +1,0 @@
-from math import sqrt
-
-
-def calculate_distance(x1, y1, x2, y2):
-    return sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
