--- operators/utils/convert_duration_to_frames.py
@@ -1,2 +1,0 @@
-def convert_duration_to_frames(context, duration):
-    return round(duration * context.scene.render.fps / context.scene.render.fps_base)
