--- operators/utils/is_ffmpeg_available.py
@@ -1,17 +1,0 @@
-import subprocess
-
-
-def is_ffmpeg_available():
-    """
-    Check if ffmpeg is installed and usable
-
-    Returns
-    -------
-    bool
-    """
-    try:
-        subprocess.call(["ffmpeg", "--help"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        return True
-
-    except OSError:
-        return False
