--- operators/utils/set_preview_range.py
@@ -1,11 +1,0 @@
-def set_preview_range(context, start, end):
-    """Sets the preview range and timeline render range"""
-    if not (start and end) and start != 0:
-        raise AttributeError("Missing start or end parameter")
-
-    scene = context.scene
-
-    scene.frame_start = start
-    scene.frame_end = end
-    scene.frame_preview_start = start
-    scene.frame_preview_end = end
